read system settings
